add edit page and layout. refactor sidebar
