+++ src/app/members/edit/page.tsx
@@ -1,0 +1,5 @@
+import React from "react"
+
+export default function MemberEditPage() {
+  return <div>MemberEditPage</div>
+}
